requirements.txt
fastapi
uvicorn[standard]
wsproto
aiohttp
azure-identity
python-dotenv
pyyaml
